Make linear solver the identity when skipped
Add an action that makes an iterative linear solve the identity operation
if no steps were performed at all, i.e. assume A=1 so Ax=b solves to x=b.
This is useful when the solver is used as preconditioner, because then we
can disable preconditioning by just not iterating the preconditioner, i.e.
by setting its number of iterations to 0 in the input file. The alternative
(when not using this action) is to keep the fields at their initial guess x=x0.