Fix PyBindings with InfoAtLink
The PyBindings lib provides a dummy definition for info_from_build
instead of defining it at link time. This commit adds the missing
include for the info_from_build function. It also moves the definition
to its own file, since it's unrelated to CharmCompatibility, and
cleans up the CMakeLists.txt for the PyBindings lib.