MockRuntimeSystem: serialize/deserialize cache contents.
This change reveals places in ActionTesting tests where the
cache contents are not properly serializable or where derived
classes have not been registered with charm.

Eventually, when MockRuntimeSystem mocks nodes and cores, we
will need to serialize/deserialize cache contents because we
will copy the contents to a separate cache per mock core.