Generalize Weno oscillation indicator to Gauss meshes
The previous code would correctly compute the indicator matrix for
GaussLobatto or Gauss points, but the caching was not designed to
handle both at the same time.

To enable Gauss meshes, this commit makes some adjustments to
documentation and assertions, it generalizes the cache to handle
both GaussLobatto and Gauss grids in the same executable (we may not
need this feature in production, but it makes testing easier), and
adds tests.