Add struct for marking DataBox tags for a reference item
A reference item refers to a part of another item (usually a tagged
container such as GlobalCache or a Variables) stored in a DataBox.
Currently these items are either tagged with a db::ComputeTag that
returns by reference, or with a db::SimpleTag that is added to the
typelist compute_with_subitems_tags of the DataBox.  This new tag
is being introduced to distinguish the tags for these items from
the tags for simple and compute items which store a type.