Update Blaze version requirements
Recent commits require blaze v3.7, but did not bump the version
requirements in CMake or in the installation docs.

This commit *only* bumps the version in these two places, but leaves
the hard work of cleaning up our blaze interface to an upcoming PR
that will update to/require blaze v3.8.

Also removes some trailing whitespace