Add Periodic boundary condition class
This class is used to mark a system-specific boundary condition as representing
a periodic boundary condition. The system-specific boundary condition needs to
inherit from the Periodic class as well as the system-specific boundary
condition base class. Domain creators then use the is_periodic function to check
if a system-specific boundary condition is marked as being periodic.