Add Jacobians and normal magnitude to MortarData
When using local time stepping on a moving mesh we need the magnitude of the
normal vector at the past times because the scheme integrates the lifted
boundary correction, and for the lifting we need the normal magnitude. When
using Gauss points we also need the face and volume determinant of the
Jacobian.