Refactor ApplyBoundaryCorrections to make LTS support easier
Local time stepping works quite differently from global time stepping. The major
difference is that with LTS the mortars are computed and lifted, then integrated
in time, and only then added to the evolved variables. In contrast, for GTS the
mortars are computed, lifted, and added to the volume time derivative, and then
the volume time derivative (which now includes the mortar terms) is integrated
in time. There is a lot of code overlap, so instead of duplicating the code,
this commit factors out what can be factored to make adding LTS support easier.