Add Subitems tags for subitems of Variables interface compute items
For interface compute items of Variables, the subitems are interface
items of the individual Tensors of the Variables.  As these are
implemented in terms of non-owning DataVectors, they cannot be
reference items, but instead are compute items.  Therefore they
require a specialization of Tags::Subitem that derives from
db::ComputeTag.