Replace DataBoxLeaf with Item in implementation of DataBox
DataBox is a tagged container that holds a set of items.  Previous to this
commit, DataBox inherited from a set of db::detail::DataBoxLeaf to manage
the items.  DataBoxLeaf presented a uniform interface for all types of
items (simple, compute, and reference) and managed them with Deferred objects,
which in turn used boost shared pointers of an abstract base class assoc_state.
This PR replaces this hierarchy of classes with three specializations of
the class template Item that directly manages the items, as well as exploiting
if constexpr (which was not available when DataBox was originally written).

In order to use the new implementation, it is necessary to wrap the subitems
of compute items in ::Tags::Subitem so the correct specialization of Item
can be used by the subitems.