Add compact elliptic DG operator
This elliptic DG operator will replace the scheme that is currently shared
between elliptic and evolution executables. By decoupling the elliptic from
the evolution scheme the new DG operator doesn't have to treat the
first-order "auxiliary" variables on the same footing as the "primal"
variables anymore. Instead, it can directly apply to the "primal" variables now,
essentially computing their second derivates, modified by the fluxes and sources
of the elliptic system and by DG boundary corrections. This significantly
reduces the size of the operator and therefore its memory usage and computational
cost for solving it. Note that this new "compact" DG operator is actually
equivalent to the first-order operator it replaces, in the sense that their
solutions are exactly the same. It's the same scheme, but reduced to solving
a smaller operator. This is facilitated by taking a Schur complement of the
first-order operator, which is possible because the auxiliary equation can
be solved for the auxiliary variables just by inverting the DG mass matrix.