Simplify code in DataBoxTag.hpp
With the recent changes in the implementation of DataBox, additional
simplifications are possible in implmentation details found in DataBoxTag.cpp

First, db::detail::item_type<Tag> can be eliminated as it should always be equivalent
to Tag::type.  This removes the ability to specify a base tag for the `return_tags`
type alias of the functor used by db::mutate::apply.  This ability was only used in
Test_DataBox and seems sketchy as I am not sure how you mutate an object without
knowing its type.

Second, since db::detail::const_item_type is no longer publically exposed and only
used in the implementation of DataBox, it can be greatly simplified.  Now it always
takes the tags list as a second template parameter.  And since all tags except base
tags are derived from a simple tag, the lookup of the type is greatly simplified.