Refactor Minmod test
Add tests for minmod_impl. This will enable increased confidence
in using minmod_impl in system-specialized minmod limiters.

Some tests that previously called the limiter are changed to now
call minmod_impl, because the lines of code being tested were in
minmod_impl and not the limiter wrapping minmod_impl.