Merge Wedge2D into Wedge
The Wedge coordinate map now works in 2D and 3D, with no change in
functionality. In particular, coordinate axes are aligned the same
as they were in the separate Wedge2D and Wedge3D implementations.
This can/should be made more consistent in the future by aligning
the radial direction with the xi coordinate-axis for both 2D and 3D
versions (see issue #2988).

Also add some more diagnostic output in tests.