Split test_creation into two functions
The ability of test_creation to take a tag was removed, and a new
test_option_tag function that takes a tag and infers the created type
was added.  This reduces the duplication of information when testing a
tag (passing both the type and tag), and makes passing metavariables
easier.