Add grid index to SegmentId
The grid index identifies an element across multiple grids. We store
the grid index (just like the block ID) in the SegmentId so that an
ElementId can just store three SegmentId, each the size of an int.
The ElementId is responsible for managing the block ID and grid index
of its segment IDs.