Compute boundary terms from package data in python tests
This is necessary to end up with a (reasonably) consistent state for the
input. The specific case where this came up is having an ASSERT in the HLLC
solver that the contact speed is between the left and right speed. This is
violated if all numbers are just random. While presumably appropriate bounds
could be set, it seems better to just feed in more "physical" random values.

In the future we may want/need to add a way of restricting the input into the
package data function, e.g. by generating random cons/prims, and setting fluxes,
etc. all consistent from that state.