Interpolation is now done using `compute_vars_to_interpolate`.
This change applies to cases where the Interpolator
ParallelComponent is used.  Interpolating without an
Interpolator ParallelComponent still uses `compute_items_on_source`
for now.

`compute_vars_to_interpolate` specifies a function that takes
as input the Variables that are sent from the Elements to the
Interpolator, and computes the set of volume Variables that are
interpolated to the InterpolationTarget.  Because
`compute_vars_to_interpolate` is a single function, it has two
advantages over the previous method (a list of ComputeItems)
 - Can use many fewer memory allocations.
 - Code is in a function, rather than a large chain
   of ComputeItems specified in the Metavariables.

Changes include:

 - intrp::Tags::VolumeVarsInfo now has vars_to_interpolate.
 - TryToInterpolate includes time-dependent Jacobian logic.
 - AH finder is now tested for finding AH in the grid frame.