Test_CylindricalEndcap: Avoid prohibited parameter choices.
Fixes #3230

The problem was that for certain (very unlikely) choices of
random numbers, the test tried to construct a map with bad
parameters.

The error was unlikely because for it to occur,
the vector from center_one to center_two (both centers are randomly
chosen in 3d) had to be nearly aligned wth the plus z-direction,
and the (randomly chosen) value of radius_one
had to be near its minimum allowed value, and the (randomly chosen)
value of radius_two had to be near its maximum allowed value, and the
(randomly chosen) value of z_plane had to be near its minimum
allowed value, and the (randomly chosen) projection point had to
be in a certain range.

Now the max and min allowed values of a few of these randomly
chosen parameters are changed to ensure that the error cannot occur.