Specify max points central in Spectral/Projection.cpp
Sometimes for testing subcell code I need one element with a large number of
subcells. The easiest way to do that is to increase the number of DG points but
the StaticCache with several ranges from say 2-50 causes compile time to
explode. Being able to temporarily locally adjust the limit makes testing some
subcell code a lot easier.