Store elliptic internal and external face data in unified map
There's not need to have separate tags for quantities on internal
and external faces. It just makes the code more complicated and
inflates the DataBox (can be quite significant when Variables are
sliced to internal and external faces). This commit switches
`domain::Tags::Interfaces<Internal/ExternalDirections>` to
`domain::Tags::Faces` in the elliptic code.