Add Schwarz smoothing to elliptic executables
Precondition the elliptic solves with a few Schwarz-smoothing
iterations. Drastically reduces the number of linear solver steps
at the cost of making each more expensive, but parallelizable.
Can be more expensive on a single core and for small problems,
but scales to multiple cores and large problems, which the
un-preconditioned solves did not (at all).

Upcoming work will add a multigrid solver, which runs the Schwarz
smoothing on a series of coarser grids. I'll also add more
subdomain preconditioners to make the Schwarz iterations faster.