Add phase bookmarks to Algorithm
This records the step of the phase where the algorithm leaves off, so that if it
returns to the same phase (e.g. from checkpointing or load balancing), it will
start from the last visited action instead of starting from the beginning again.

This allows us to relax the restriction of `ExecutePhaseChange` needing to be
the last action in the list, and provides more flexibility for various
interoperability with phase changes.