Move time-stepper tags from global cache to databox
It is easier to support separate components stepping differently (e.g. GH and
CCE) when the relevant objects used by the time-stepping infrastructure do
not need to have unique tags.

This also has the philosophical advantage that when a component does not use
a given collection of time-stepper based tags, they don't appear in
that component's DataBox.

Most time-stepper objects are small, so this should negligibly impact the
memory usage for the array components.