Update tolerance for radial mesh velocity in Bjorhus class
When we have grids that are tangentially moving at the outer
boundary, e.g. uniformly rotating spheres, the normal mesh
velocity may get a floating-point error level outgoing
velocity. As we do not want this to throw the error that is
supposed to be thrown when the user genuinely uses an outward
moving outer boundary, we update the tolerance for triggering
of that error.