Clear LibraryVersions.txt file at start of CMake
Previously, the version file was cleared at end of CMake, which would
erase any information about library versions written into the file
during the CMake configuration. By clearing the file at the start of
CMake, we keep all information from the configuration.

To streamline how things are written into the file, we use
PrintUsefulCMakeInfo to append build info to LibraryVersions. Since the
info being written to the LibraryVersions.txt file and the info being
printed to the command-line were nearly identical, just do both from the
same file. This is done at the end of CMake, and ensures the build info
being written into LibraryVersions.txt is up-to-date.