Remove unnecessary headers in AnalyticData/CurvedWaveEqn
These headers were intended to hide the includes for
ScalarWaveGr<SW, Bkground> for different choices of SW and
Bkground. However, even the unit tests of these classes do
not use this compound header and instead include
ScalarWaveGr, SW, Bkground individually. This is the more
natural choice anyway, so this commit removes the
unnecessary compound headers.