Add output operators for DataBox and AlgorithmImpl
These are useful for debugging and exploring.
For DataBox, adds two print functions, one that prints the expanded
type aliases, the other prints the items (but not subitems) held.
For AlgorithmImp, adds four print functions, one for the expanded type
aliases, the second for the state, the third for the contents of the
inbox, and the fourth for the contents of the DataBox.