Add option to apply Persson TCI to GRMHD TildeB
So far this has only been necessary for cases where the magnetic field has a
discontinuity but the fluid variables are smooth. Specifically, the magnetic
loop advection problem. Magnetized TOV stars seem to do okay without this, but a
careful comparison on whether the Persson TCI should be applied to TildeB in
the magnetized neutron star case should be done.