Fix PiecewisePolynomial results at update times
It is allowed to update at the last valid time for evaluating the
function.  An evaluation at this time must not reflect the effects of
the update because it could have been previously evaluated before the
update was performed.

Also instantiate more cases to make testing easier and for potential
later use.