Add pup function to circumvent charm bug for map with comparator
Adds a function `pup_override` to use for `std::map`s with custom comparators,
because currently a charm bug causes an invalid `reserve` call for such
containers.
The function is intentionally chosen to not directly interfere with the charm
pup override so that the charm version of the function can be easily used or
tested once the bug is fixed.