Remove incorrect assertion in component mocking
The assertion of the existence of a component_being_mocked type alias
was being performed too aggressively.  In tests where some components
were mocking and others were not, it would sometimes give confusing
error messages about specific non-mocking components not mocking
something it was looking for (even when an appropriate mocking
component existed).