Support per-core reductions
The reduction observer now supports writing the intermediate stage of
the reduction over just the core that the element is currently on. This
can be useful e.g. to measure performance metric to assess
load-balancing such as the number of elements or grid points on
each core. The data will be written in one H5 file per _node_ prefixed
with the reduction file name, in a `Core{core_id}` subfile. For example,
when running on 2 nodes with 2 cores each you will end up with
`Reductions0.h5` containing `/Core0/{subfile_name}.dat`
and `/Core1/{subfile_name}.dat`, and `Reductions1.h5` containing
`/Core2/{subfile_name}.dat` and `/Core3/{subfile_name}.dat`.