Support time indices for some TensorExpressions
The commit includes the minimal implementation needed to get time indices working with TensorExpressions at all. The result of these changes is that time indices can be used to write tensor equations with TensorExpressions that evaluate a single RHS tensor to a LHS tensor (including transposing indices) or that contain contractions and/or products. In short, the changes here do not yet support writing equations that contain addition, subtraction, or taking the square root of an expression that evaluates to a scalar. With these changes, concrete time indices can be used for spacetime indices in either or both the RHS and LHS tensors in an equation.