Add virtual functions to FunctionOfTime base
To avoid having to cast whenever we need to update or reset the
expiration time of a FunctionOfTime, this adds `update` and
`reset_expiration_time` as virtual functions of the FunctionOfTime
base class. The default of these virtual functions is to ERROR
because by defualt FunctionsOfTime shouldn't be able to update.