Use ObservationBox to reduce ScalarWave memory usage
Stop storing the spatial derivatives, one and two index constraints, and
pointwise L2 norms of the one and two index constraints in the DataBox. They are
now added as compute tags to the ObservationBox when running events.

In the future we can remove ObserveErrorNorms by using the ErrorCompute tag and
ObserveNorms.