Use pow instead of log in Persson TCI
Evaluation of the TCI criteria is switched into an equivalent form in which
log() function is removed from both sides and the division by
l2Norm(component) is moved. In this way zero_cutoff (which was used for a
safety check of computing log() for zero or a very small value) can be removed
and also division by zero can be avoided when the component norm is zero.