Interpolate Ah quantities to 2 frames if desired.
Previously, if we found the horizon in the grid frame, we could
run observers only in the grid frame, because the metric quantities
were computed on the horizon only in the grid frame.

Now, if we find the horizon in the grid frame, we can optionally
interpolate (and hence observe) quantities in the inertial frame as well
as in the grid frame.  Technically this means that we are now
allowed to specify quantities in both frames in DestTagList in
ComputeHorizonVolumeQuantities.hpp.

Note that if we don't want quantities in both frames, then those
quantities are not computed (except for any intermediate quantities
that are required to compute the desired output in the desired
frame).  And the computations are still done with minimal
memory allocations.