Add FoT tag initialized with control systems
This tag takes some `control_system::OptionHolder`s as option_tags
for its `create_from_options` member in addition to the domain
creator and initial time step. It is similar to the tag in
`domain::Tags::FunctionsOfTimeInitialize` except that this new tag
depends on the control system, so it must be located outside the
Domain library.