Fix a bug with instantiating multiple TOV solutions
The radial TOV solution was kept in a static variable of the
RelEuler::TovStar class. I may be missing something, but this seems
like a serious bug to me. If two TOV solutions with different parameters
are created, the second will inconsistently use the solution of the
first. This commit fixes the issue by storing the radial solution as a
member variable.