Improve error message in file comparison-type tests
Previously if the subfile wasn't in the HDF5 file, maybe because of a typo, it
was not clear _why_ the test failed. Now we report the subfile we are trying to
find, that the check failed because the subfile wasn't there, and what the
available subfiles are in all the HDF5 files that matched the glob.