InterpolateWithoutInterpComp: Use compute_vars_to_interpolate.
Previously it created a temporary DataBox and evaluated ComputeItems.
Now uses a mutator function compute_vars_to_interpolate, just like
interpolation with an Interpolator ParallelComponent.

The compute_vars_to_interpolate can do frame transformations if it
chooses, just as for interpolation with an Interpolator
ParallelComponent.

Without an Interpolator ParallelComponent, there is still the
restriction that the interpolation
points must be constant-in-time in some frame.
But that frame can be anything.