Blaze: enable BLAS kernels
Blaze uses BLAS kernels for max performance; in particular the
BLAS dense matrix multiplication kernel is faster than Blaze's fallback
according to their docs:
https://bitbucket.org/blaze-lib/blaze/wiki/Configuration%20and%20Installation#!step-2-configuration
Therefore, we follow the Blaze recommendation and make sure BLAS
is enabled and linked.

Also make sure LAPACK is linked, which is a dependency of Blaze.