Remove some annoying StrahlkorperTag aliases.
These were aliases that caused confusion instead of helping,
because the types they referred to are short and clear.

The aliases for Jacobian, InvJacobian, and InvHessian still
remain, because those are complicated because of the Pfaffian
derivatives and the fact that some of the indices are spherical
and others are cartesian.