XCTS binary: superpose matter without Gaussian falloff
Matter sources are superposed without the window functions. They are of
limited use anyway, because in a binary setting they don't take the
gravitational influence of the other body into account. Therefore, the matter
sources should typically be solved-for alongside the gravity sector to impose
conditions such as hydrostatic equilibrium. For scenarios where we just want
to superpose the isolated matter solutions and compute the resulting gravity,
the matter sources are simply added.