Optionally bootstrap missing Py dependencies in build dir
When the optional `BOOTSTRAP_PY_DEPS` CMake flag is enabled, missing
Python dependencies are installed to the build directory.
This is an alternative if the user doesn't want to bother installing
the dependencies themselves, e.g. on a cluster. Packages like
h5py, numpy and scipy can/should still be installed by the user
to make sure they use the correct HDF5, BLAS, etc., but the new
`BOOTSTRAP_PY_DEPS` flag makes it easy to install pure-Python packages.