Add `PY_DEV_MODE` CMake flag
The flag enables development mode for the Python package, meaning
that Python files are symlinked rather than copied to the build directory.
Allows to edit and test Python code much easier, in particular when it uses
compiled Python bindings, but doesn't replace CMake placeholders
in the Python code such as the project version.