Consolidate mutation of global cache into one function
In order to avoid having multiple mutate functions for the action
testing framework and for charm-aware executables, use only a single
Parallel::mutate function that takes a reference to the local cache.
Then, if Parallel::mutate was called in an executable, it will call
mutate on the proxy, but if it was called in a test, it will just
mutate the local cache.

This came up because previously, our testing framework could not handle
calls to Parallel::mutate that happen in source code (we didn't actually
have any so everything was ok).