Omit BBH domain cube-to-sphere transition
The cube-to-sphere transition is not needed when the layer of frustums
is already spherical, as added by #3598.
This commit omits the 10 blocks that do the cube-to-sphere transition
when the frustums are already spherical.
The old behavior is still supported by just setting the
frustum sphericity to 0. We may remove it at a later time, once we
are satisfied the bulged frustums work sufficiently well.