Always select same ElementIds in element_logical_coordinates
For points on Element boundaries shared within the same Block, the selected
ElementId will be the one on the upper side in each shared dimension,
instead of the first of the input ElementIds that share the point.

Update an existing test as the selected Element changed.

Test that the same ElementIds are selected by randomly shuffling the input
ElementIds.