YlmSpherepack: Replace mutable non-dynamic storage.
This addresses part of #3929 but not all of it.

YlmSpherepack has 2 types of internal storage:
1. Quantities computed only once and reused often, but forever constant.
2. Temporary buffers that are overwritten frequently.

This commit changes 1. and does not address 2.

Previously, the temporary quantities in 1. were computed using
lazy evaluation: if a member function that needed them was
called, then they were computed and then forever held fixed.
This was accomplished via a mutable member variable that holds the
storage.  Now, the temporary quantities in 1. are all computed
in the YlmSpherepack constructor, and are never changed.  The
corresponding member variable is no longer mutable.  In the
constructor, certain temporary quantities are allocated on the heap
on the fly and then thrown away after use, but because this happens
only in the constructor we don't worry about these allocations.