Handle late calls to InterpolatorReceiveVolumeData.
There can be elements that are unneeded for interpolation,
but yet these elements still call InterpolatorReceiveVolumeData.
(This is necessary, because e.g. the 7th apparent-horizon iteration
might move the surface into a new element).
However, an unneeded element may call InterpolatorReceiveVolumeData
after the interpolation is complete; in that case we now ignore that
call to InterpolatorReceiveVolumeData.