intrp_detail::block_logical_coords takes coords as input.
So now you can pass in the coords in the frame that you want the output.

The old overloads of intrp_detail::block_logical_coords are still
there and are still used; those compute the coords internally
by calling the InterpolationTarget.

The new overload is used for the case without an Interpolator
ParallelComponent, where the coords are stored.