Remove load balancing test
The test doesn't work for any version of Charm++ up to v7.0.0 because of
a Charm++ bug.  It triggered very rarely with v6.10.2, but seems to be
occuring more frequently on the clang-13 build.  The removed test just
added load-balancing to another test which was difficult to debug when
load-balancing failed.

Charm++ has fixed the issue and I have written a better test which I will
add when we next update the Charm++ version we use.