Add size control logic and test.
This is done via an abstract base class SizeControlState
that has derived classes for each concrete state of the size
control system.

The logic (and control errors) are added in such a way that the
code can be tested independently.  There is (intentionally) no
dependence on Strahlkorpers, FunctionsOfTime, ZeroCrossingPredictors,
or the control system framework itself.

The thing that calls the SizeControlState member functions
(and thus needs to know about Strahlkorpers, FunctionsOfTime, etc so
that it can compute the arguments to those member functions)
will be added in a later PR.