Remove Parallel::Phase::RegisterWithObserver
In elliptic and evolution executables, replace with Parallel::Phase::Register
to make the name of the phase consistent.

In some test executables, replace with Parallel::Phase::Testing, as these
executables were already using the Register phase.  In these tests the
phase names have no meaning, but need to be distinct.