Update dev container
Previously, cmake only found Ninja to build with. Now it finds make instead.

Since the docker container is run as its own user "docker", git complained that the
spectre repo was owned by a different user (the actual user). This caused cmake to
fail to run. This ensures that the spectre repo is a "safe" directory.