Replace spin_vector function with new version.
The old spin_vector used a single frame, and assumes that
1) the Strahlkorper is in that frame, 2) that the metric quantities
are in that frame, and 3) that the metric quantities are evaluated
at the collocation points of the Strahlkorper.

Those assumptions turn out to be difficult and expensive
to achieve in practice when observing the spin_vector in a frame different
than the frame that the Strahlkorper is computed in.
Especially 3) would require an additional interpolation.
See issue #4132 for details.

The new spin_vector uses two frames.  It also assumes 1), 2), and
3) for one of those frames ('MetricDataFrame'), but it also takes as
an argument the Cartesian coordinates of the Strahlkorper in a
different frame ('MeasurementFrame'), evaluated on the MetricDataFrame
collocation points of the Strahlkorper.  The resulting spin vector
is returned in MeasurementFrame.

Now when computing spin_vector
in a frame different than the frame that the Strahlkorper is computed
in, there is no need to transform the Strahlkorper into a different
frame, and there is no need for extra interpolation of metric
quantities.