UniformCylindricalEndcap: check more bracket-fail cases.
In the inverse function, if the root isn't bracketed, there is
a check that the bracketing failure is caused by roundoff.
This check was done only if the function value was close to
zero.

This commit changes what 'close' means, and does the check much
more often.  Also this commit puts the value of 'close' into
a constexpr variable instead of hardcoding it into two 'if'
statements.