ObserveCenters: use Euclidean area element.
The Euclidean area element is probably more appropriate, and
matches better the behavior of other algorithms to compute
the Strahlkorper center in a different frame (such as
computing a full Strahlkorper in a different frame and then
computing the center; that process is independent of the metric).

Also fixes a bug in how centers were computed, and adds Tags
declarations that were missing.