Make running EventsAndDenseTriggers deterministic
Previously the order depended on hashes of memory addresses.

Additionally, triggers are now always tested in input-file order.
Previously, the order that triggers were checked in would vary in a
deterministic but not easily predictable manner throughout the
evolution.

This incurs an asymptotic runtime penalty because the vector of
triggers is being looped over repeatedly, but unless the number of
triggers becomes very large this effect should be negligible.