Raise SIGTRAP in Blaze when in debug mode
Since the option parser uses exceptions quite a bit, using 'catch throw' in GDB
is not the best experience. Blaze has support for custom error handling, so we
will now raise SIGTRAP before throwing. This is analogous to what we do with our
own errors. It also disables error checks in Blaze in release mode, consistent
with how we disable all error checks in SpECTRE in release mode.