Remove Initialization::Action::RemoveOptionsAndTerminatePhase
Also remove type alias initialization_items_to_keep.

With the the change to a single DataBox in a DistributedObject, an item
was no longer removed, but reset to a default initialized value.
As only a few items are currently removed (and they are fundamental types),
it makes more sense to explicitly mutate items if desired, instead of
relying on scattered type aliases.