Have SpECTRE control system act more like SpECs control system
The way it acts more like SpECs is that all control systems use
the same measurements and use a specific number of measurements
before it updates the functions of time.

It differs from SpECs in that the Nth derivative of the
functions of time aren't changed at the update time, but rather
one measurement later. This is done to avoid extra global syncs.