New epsilon in ZeroCrossingPredictor test.
The scale is now set to unity, since the data points are
spread over a range of order unity.

Previously the test failed on crossing values near zero
because the epsilon wanted many many digits in that case.

Also, increased the value of epsilon slightly, because
the test also failed occasionally for larger values.