UniformCylindricalSide: Allow extra use case.
In the case where z_plane_plus_one=z_plane_plus_two or
z_plane_minus_one=z_plane_minus_two, no longer demand
that z_plane_plus_one and z_plane_minus_one
are on the opposite sides of the equator of sphere_one.

This is important for the case of the BBH cylindrical domain
where there are extra cylindrical outer shells (what SpEC
calls "SphereC"s).

This commit does not change the map at all: it changes only the
set of allowed input parameters, the parameters that are used
in the test, and the documentation.