CylindricalBinaryCompactObject: use new maps.
Now uses UniformCylindricalEndcap, UniformCylindricalSide,
and UniformCylindricalFlatEndcap instead of the old maps
CylindricalEndcap, CylindricalSide, CylindricalFlatEndcap.

Also now has the option to add an extra outer cylindered-sphere
shell, and to add an extra inner cylindered-sphere shell
for either of the excised regions.

Does not yet handle time-dependent maps in the way we
want it to, but that can be the next PR.