Add new receive_data entry method to DistributedObject
This one takes a Charm++ message (pointer)  and inserts it into the
inbox.

Also, a corresponding Parallel::receive_data function was added. This
can only be used on Array components at the moment.

You also need to properly register messages with charm or you'll
get incorrect behavior so there's also code added to handle that
automatically.

The parallel test for this can be run on one or two charm nodes.