Add TciDecision tag and a TciStatusCompute tag
We want to start sending the TciDecision to our neighbors in order to support
adding a halo layer of FD elements around troubled elements. The
TciStatusCompute tag allows us to only store the TCI decision on a grid during
observation, reducing memory usage.

A followup commit will add the tags to the executables.