Add ability to override observation value to ObserveFields
This is necessary when dumping volume data during cleanup after a failure since
elements are not guaranteed to be at the same time value. E.g. one element
updates its time, then another does a boundary correction and errors before
updating the time. We want all the volume data to be dumped at the same "time"
so it's easier to visualize.