Infer output buffer tags in partial_derivatives
This allows the compiler to infer the template parameters in
the return-by-reference overloads of `partial_derivatives`.
It also relaxes the restriction that the output buffer must have
`::Tags::deriv` prefixes. This is useful to compute derivatives
directly into output buffers, e.g. for variables that are
represented by another tag (like Phi = deriv(Psi)) or for
variables augmented by more computations (like adding a
Christoffel-symbol term for the covariant derivative) or for
prefixed variables where the output buffer ignores the prefixes.