Factor out list of tags initialized in a DistributedObject
The distributed object tags are those corresponding to items
in the DataBox of a DistributedObject that are initialized by
the constructors of a DistributedObject.  This list needs to be
available to amr::Actions::AdjustDomain in order to ensure that
adaptive mesh refinement initializes all items in the DataBox
when new elements are created.