Add SetInterpolators
This sets interpolators for going from DG to FD and FD to FD at block boundaries
where the grid coordinates are not smooth. These are used to set ghost zones
sent for reconstruction. A separate part of the code will need to be responsible
for updating the interpolants should we allow p-refinement with DG-FD.