Add ComovingCharSpeedDerivative
This is a function that computes the derivative of the
comoving characteristic speed with respect to the map parameter.
The sign of this derivative is used in the control system
for state DeltaRDriftInward. Basically, if a particular state
will cause the comoving characteristic speed to decrease, then
we don't enter that state.

Note that this is not exactly the same quantity computed in SpEC
(but I think that SpEC should be computing this quantity and
not what it currently computes).