Remove layer of autogenerated code from executables
Define executables in a cpp file instead of auto-generating
that cpp file. This removes the need for a `Fwd.hpp` file,
removes static variables with special names, and makes it
(IMO) much easier to understand how an executable is
compiled.

As a side product this PR also makes executable headers
compatible with clang-tidy because they don't contain
undefined symbols like `metavariables` anymore.