Fix interpolation framework when there are no valid target points
Prior to this change, if there were no valid target points, the interpolation
target never received a subsequent message, which meant no cleanup was done,
and the current temporal_id kept waiting for a message.  For sequential targets
this leads to a growing list of pending temporal_ids and accumulated volume data
waiting to be interpolated.

Now the interpolation target is sent a message, and will continue to do whatever
it would have done if there was a subset of target points that are invalid.

Added additional tests to some interpolation callbacks that would have failed
prior to this change when there are no valid target points.