Simplify elliptic DG formulation
Remove the notion of "auxiliary variables". Just always choose the
partial derivatives of the system variables as auxiliary variables.
This doesn't change Poisson and Puncture systems at all,
but Elasticity and XCTS systems change at
truncation error level.

This simplification helps to reformulate the DG scheme in a next PR.
It also makes it easier to implement elliptic systems altogether
and improves performance by ~20% because it
eliminates Kronecker-deltas in the formulation.