Add Post-Newtonian Conformal part to conformal metric in WavyBBH.
The class WavyBBH was changed so that it takes as input parameters:
 - both masses of bh;
 - both x positions;
 - both y momentums.
The methods to calculate the distance to each black hole were added. (Radius_... Tags)
The methods to calculate the normal vector pointing to each bh were added. (Normal_... Tags)
The conformal part of the Post-Newtonian metric was added to the conformal_metric calculation.