Add base methods compute radiative term in WavyBBH class
The methods to compute the radiative term of the conformal metric where added.
The radiative term is now a new Tag.
There is 1 method to add this radiative term to the conformal metric
(multiplied by an atenuation function - input parameter FatPar).
There are 4 methods to add the following terms to the radiative term:
 - near zone term
 - present time term
 - past (retarded) time term
 - integral term
For now the radiative term is set to zero.
The code compiles and it can be run with SolveXcts, input file WavyBinaryBlackHole.yaml.
The background data is now being added to the output in the first iteration.