Add TraceExtrincicCurvature computation to WavyBBH
The trace of the extrisic curvature in Post-Newtonian orders is explicitly computed.
The tags for the derivatives of the radius_left(right) and of the functions 1/radius_left(right) were added.
The derivatives are explicitly computed.
Changes in some of the parameters of the input file were done, the simulation domain is now smaller.