Use valid values in test
Choose a valid step size for the initial time, and start in a slab
number consistent with the other chosen value.

Also instead of going to the end of the slab and backing up a slab,
just go to the start directly.  This appears to have evolved as the
tested code changed, so I assume it did something saner originally.