Store penalty factor on mortars
The penalty factor is constant throughout the solve. Remove the need
to communicate data to compute it, and avoid some mortar
projections, by storing the penalty factor on mortars.

Note: this changes the penalty factor on nonconforming boundaries.
Before, the perpendicular element size on both sides of an
element face was projected to the mortar and communicated. Now, the
perpendicular element size is computed directly on the mortar. This
should make it more accurate (since the mortar is always higher
resolution than the faces).