Elliptic DG: communicate only fields and fluxes
Before, parts of the boundary corrections were already computed
by the element that sends the data. However, this is incompatible with
an upcomping small change to the DG scheme to handle elasticity
with discontinuous materials. I also find it easier to understand
the code if only the fields and fluxes are communicated, and it
reduces the amount of communicated data.

Note: this changes which quantities are projected to and from mortars
at nonconforming boundaries.
Before, fluxes and normals were applied to the fields for the
auxiliary numerical flux and the jump term, then projected and
communicated.
Now, the fields are projected and communicated directly.
Averages and jumps are computed on the mortar, projected back to the
face, and then fluxes and normals are applied.