Add static assert to partial derivative
There used to be a check that the tags actually worked out, but that was broken
when the tags were no longer required to be Tags::deriv. Check that the
underlying tensor types are the same is the best check I can think of to
guarantee that you are doing what you want.