Fix Puncture AMR criterion on element boundaries
Before, if the puncture is on a shared element boundary, only one of the abutting
elements was refined. This was because the `{block,element}_logical_coordinates`
functions disambiguate points on element boundaries so they always map to one
unique element. This is not the expected behavior for the AMR criterion.
Now, all abutting elements are refined.