Introduce MortarDataHolder and use it
A MortarDataHolder holds two MortarData representing the data on the
local and remote sides of a mortar between two elements.  This new
class will be used for global time-stepping in a similar way to how
the BoundaryHistory is used for local time-stepping.  This will allow
the MortarData class to be simplified so that it represents the data
on only one side of the mortar instead of both sides.